Stats: Minor fixes
- Allocate matrices as part of Stats::Permtest::precompute_default_permutation() rather than requiring pre-allocation.
- Fix various signed - unsigned integer comparison warnings.
- Use .eval() calls to evaluate Eigen matrix results prior to invoking Math::pinv(); this appears to be required for older versions of either Eigen or GCC.
- Fix final permutation not being assessed.

+++ cmd/mrclusterstats.cpp
@@ -134,7 +134,7 @@ class SubjectVoxelImport : public SubjectDataImportBase
     void operator() (matrix_type::RowXpr row) const override
     {
       assert (v2v);
-      assert (row.size() == size());
+      assert (size_t(row.size()) == size());
       Image<float> temp (data); // For thread-safety
       for (size_t i = 0; i != size(); ++i) {
         assign_pos_of ((*v2v)[i]).to (temp);
